+++ modules/tinymce/src/core/main/ts/EditorSettings.ts
@@ -37,7 +37,6 @@ const deviceDetection = PlatformDetection.detect().deviceType;
 const isTouch = deviceDetection.isTouch();
 const isPhone = deviceDetection.isPhone();
 const isTablet = deviceDetection.isTablet();
-const legacyMobilePlugins = [ 'lists', 'autolink', 'autosave' ];
 const defaultTouchSettings: RawEditorSettings = {
   table_grid: false,          // Table grid relies on hover, which isn't available so use the dialog instead
   object_resizing: false,     // No nice way to do object resizing at this stage
@@ -50,11 +49,6 @@ const normalizePlugins = (plugins: string | string[]) => {
   return Arr.filter(trimmedPlugins, (item) => {
     return item.length > 0;
   });
-};
-
-// Filter out plugins for the legacy mobile theme
-const filterLegacyMobilePlugins = (plugins: string[]) => {
-  return Arr.filter(plugins, Fun.curry(Arr.contains, legacyMobilePlugins));
 };
 
 const extractSections = (keys, settings) => {
@@ -73,11 +67,6 @@ const getSection = (sectionResult: SectionResult, name: string, defaults: Partia
 
 const hasSection = (sectionResult: SectionResult, name: string) => {
   return Obj.has(sectionResult.sections(), name);
-};
-
-const isSectionTheme = (sectionResult: SectionResult, name: string, theme: string) => {
-  const section = sectionResult.sections();
-  return hasSection(sectionResult, name) && section[name].theme === theme;
 };
 
 const getSectionConfig = (sectionResult: SectionResult, name: string) => {
@@ -163,11 +152,8 @@ const combinePlugins = (forcedPlugins: string[], plugins: string[]): string[] =>
 };
 
 const getPlatformPlugins = (isMobileDevice: boolean, sectionResult: SectionResult, desktopPlugins: string[], mobilePlugins: string[]): string[] => {
-  // is a mobile device with mobile theme
-  if (isMobileDevice && isSectionTheme(sectionResult, 'mobile', 'mobile')) {
-    return filterLegacyMobilePlugins(mobilePlugins);
   // is a mobile device with any mobile settings
-  } else if (isMobileDevice && hasSection(sectionResult, 'mobile')) {
+  if (isMobileDevice && hasSection(sectionResult, 'mobile')) {
     return mobilePlugins;
   // is desktop
   } else {